[@svelteui/core]: styles directory file changes & SvelteUIProvider improvements
* [@svelteui/core]: styles directory structure change

* [@svelteui/core]: created merge-theme util

* [@svelteui/core]: SvelteUIProvider major changes

* [@svelteui/core]: svelteUIProvider abstractions

* [@svelteui/core]: moved SvelteUITheme interface out of the engine dir into the theme/types dir

* [@svelteui/core]: fns file structure changes
