--- packages/svelteui-core/src/lib/styles/SvelteUIProvider/index.ts
@@ -1,5 +1,0 @@
-export { default as SvelteUIProvider } from './SvelteUIProvider.svelte';
-export { config } from './svelteui.config';
-export { colorScheme } from './svelteui.stores';
-export { useSvelteUIDefaultProps, useSvelteUIThemeContext } from './svelteui.provider';
-export { useSvelteUITheme } from './default-theme';
