--- packages/svelteui-core/src/lib/styles/SvelteUIProvider/svelteui.config.ts
@@ -1,6 +1,0 @@
-import type { SvelteUIProviderProps as $$SvelteUIProviderProps } from '../types';
-
-export const config: $$SvelteUIProviderProps['config'] = {
-	light: { bg: 'White', color: 'Black' },
-	dark: { bg: '#1A1B1E', color: '#C1C2C5' }
-};
