+++ packages/svelteui-core/src/lib/styles/engine/index.ts
@@ -1,4 +1,4 @@
-export type { Selectors, SvelteUITheme, SvelteUIThemeOverride } from './types';
+export * from './types';
 
 export { createStyles } from './create-styles.js';
 export { cssFactory } from './css.js';
