+++ packages/svelteui-core/src/lib/styles/engine/types/index.ts
@@ -1,2 +1,1 @@
 export type { Selectors } from './Selectors';
-export type { SvelteUITheme, SvelteUIThemeOverride } from './SvelteUITheme';
