--- packages/svelteui-core/src/lib/styles/functions/fns/radius/radius.ts
@@ -1,12 +1,0 @@
-import { theme as _theme } from '../../../index';
-import type { SvelteUINumberSize } from '../../../types';
-
-export function radius(radii?: SvelteUINumberSize): number | string {
-	const theme = _theme;
-
-	if (typeof radii === 'number') {
-		return radii;
-	}
-
-	return theme.radii[radii].value;
-}
