+++ packages/svelteui-core/src/lib/styles/theme/SvelteUIProvider/index.ts
@@ -1,0 +1,15 @@
+export { default as SvelteUIProvider } from './SvelteUIProvider.svelte';
+export { useSvelteUITheme } from './default-theme';
+export { colorScheme } from './svelteui.stores';
+export {
+	key,
+	ssrStyles,
+	useSvelteUIDefaultProps,
+	useSvelteUIThemeContext
+} from './svelteui.provider';
+export type {
+	ProviderStyles,
+	SvelteUIDefaultProps,
+	SvelteUIProviderContextType,
+	SvelteUIProviderProps
+} from './svelteui.provider';
