+++ packages/svelteui-core/src/lib/styles/theme/default-colors.ts
