+++ packages/svelteui-core/src/lib/styles/theme/functions/attach-functions.ts
@@ -1,0 +1,1 @@
+export {};
