+++ packages/svelteui-core/src/lib/styles/theme/index.ts
@@ -1,0 +1,6 @@
+export * from './SvelteUIProvider/index.js';
+export * from './functions/index.js';
+export * from './utils/index.js';
+export * from './types';
+
+export { colors } from './default-colors';
