+++ packages/svelteui-core/src/lib/styles/theme/types/SvelteUIColor.ts
