+++ packages/svelteui-core/src/lib/styles/theme/types/SvelteUIGradient.ts
