+++ packages/svelteui-core/src/lib/styles/theme/types/SvelteUIMargin.ts
