+++ packages/svelteui-core/src/lib/styles/theme/types/SvelteUISize.ts
