+++ packages/svelteui-core/src/lib/styles/theme/types/index.ts
@@ -1,4 +1,3 @@
-export type { SvelteUIProviderProps } from './SvelteUIProvider';
 export type { DeepPartial } from './DeepPartial';
 export type { DefaultProps } from './DefaultProps';
 export type { ColorScheme, DefaultBackground, DefaultColor } from './ColorScheme';
@@ -8,12 +7,17 @@ export type { SvelteUIMargins, SvelteUIMargin } from './SvelteUIMargin';
 export type { SvelteUIShadow } from './SvelteUIShadow';
 export type { SvelteUINumberSize, SvelteUISizes, SvelteUISize } from './SvelteUISize';
 export type {
-	HeadingStyle,
-	Override,
-	LoaderType,
+	Fn,
 	CSS,
-	VariantThemeFunction,
-	Transition
+	Fns,
+	Override,
+	Transition,
+	LoaderType,
+	HeadingStyle,
+	DefaultTheme,
+	SvelteUITheme,
+	SvelteUIThemeOverride,
+	VariantThemeFunction
 } from './SvelteUITheme';
 export type {
 	SvelteUIFontSizes,
