+++ packages/svelteui-core/src/lib/styles/theme/utils/index.ts
@@ -2,3 +2,4 @@ export { toRgba } from './to-rgba/to-rgba.js';
 export { rgba } from './rgba/rgba.js';
 export { vFunc } from './get-variant-theme/get-variant-theme.js';
 export { randomID } from './random-id/random-id.js';
+export { mergeTheme } from './merge-theme/merge-theme.js';
